Moved aws-sdk to a dev dependency
This requires the user of this library to pass in the dynamoDBClient option.

This prevents overwriting the aws-sdk in lambda functions, and drastically reduces the deployment bundle size of lambda functions that require this dependency.

Also, removed winston because it’s overkill for logging a few error messages.

This isn’t meant to be a PR, just needed to change these things to get it to work in our lambda functions.